### 2018-05-30 bihealth/sodar-core#5 @mikkonie
Not critical, as this occurs rarely and is just a cosmetic oddity.
