### 2018-08-27 bihealth/sodar-core#3 @mikkonie
* Model for remote sites
    * Title, description, URL, token
* Admins have to allow/disallow remote sites (UI)
    * Per-project rights for projects (non-inheriting right now)
* Different levels for remote access
    * Project exists
    * See title and description
    * See roles
    * Sync projects and roles
