### 2018-09-12 bihealth/sodar-core#3 @mikkonie
Moved to sodar_core and updated.
