### 2018-09-18 bihealth/sodar-core#4 @mikkonie
mentioned in commit 8711e4cd2b21c807691fe1c6ba28b74d84558bc2
