### 2018-09-18 bihealth/sodar-core#4 @mikkonie
Done.
