### 2018-09-20 bihealth/sodar-core#3 @mikkonie
mentioned in commit 89b786709a05600e2eb4b5689a77554040dfcb45
