### 2018-09-21 bihealth/sodar-core#3 @mikkonie
mentioned in commit b554ac313faaa2e58271f531ae83fce182b4dd70
