### 2018-09-25 bihealth/sodar-core#3 @mikkonie
mentioned in commit ee71810548ce55eb211436627307a157828308a3
