### 2018-09-26 bihealth/sodar-core#3 @mikkonie
mentioned in commit f1d9257c35ec3f7634849429dc61914edf9ce415
