### 2018-09-26 bihealth/sodar-core#3 @mikkonie
mentioned in commit bce46bdf1d8c21ee1ad4e90baeaccaf57ac79888
