### 2018-09-26 bihealth/sodar-core#3 @mikkonie
mentioned in commit 13a373ecb23a9afeb78a10969d7aaadabdea0844
