### 2018-09-27 bihealth/sodar-core#3 @mikkonie
mentioned in commit dd0a80bbc922aa1c9e47b8974914d0b6a8f41101
