### 2018-09-27 bihealth/sodar-core#3 @mikkonie
mentioned in commit 9fe925ae5d7b05672941736ad2706e2ce160b59f
