### 2018-09-27 bihealth/sodar-core#3 @mikkonie
mentioned in commit eae53ad88d8a7f9322b6ec79362bdbe28ffc8070
