### 2018-09-27 bihealth/sodar-core#3 @mikkonie
mentioned in commit edc0de25823c10b3a6cd9ec40058b076cc5ea9d2
