### 2018-09-28 bihealth/sodar-core#3 @mikkonie
mentioned in commit daff1c1bd2adf0c1fd4db15212ec76c2c70f521a
