### 2018-09-28 bihealth/sodar-core#3 @mikkonie
mentioned in commit 9647a68b9ea5b5ad9aff1c9808b5fbe68f00a03e
