### 2018-10-01 bihealth/sodar-core#3 @mikkonie
mentioned in commit e852b19423f90042b3d711a10dd7d363b7391293
