### 2018-10-02 bihealth/sodar-core#3 @mikkonie
Ran into a few things we didn't think about or discuss while working on this:

- Users have to be added too, in case a project has roles for users which don't exist in the target site
    * This is fine for LDAP users since they are unique, however local users (such as admins) may actually refer to different people even if the user name is the same
    * This is a problem especially if a category/project is owned by a local user
    * Current solution: if a project is owned by a non-LDAP user, add the owner role to the admin user initiating the project sync and do not sync any other non-LDAP roles. **TBD:** Is this OK?
