### 2018-10-02 bihealth/sodar-core#3 @mikkonie
mentioned in commit 65e9ce33febbec6d1c7761f5b3c85755c5a8a8df
