### 2018-10-04 bihealth/sodar-core#3 @mikkonie
mentioned in commit f08248fd9ff6d3d46c3c37e98717a7115ddd09f3
