### 2018-10-04 bihealth/sodar-core#3 @mikkonie
mentioned in commit a449e587fd14165b2770ba1d20edba8f022eb93a
