### 2018-10-05 bihealth/sodar-core#3 @mikkonie
mentioned in commit 439895d4e69d3d57c4211a6e000197113c043fe0
