### 2018-10-08 bihealth/sodar-core#3 @mikkonie
mentioned in commit 233aeb5521fade72ddd2c43687d060c4546127b1
