### 2018-10-08 bihealth/sodar-core#3 @mikkonie
mentioned in commit dc1240816fcf3b7c64f2f697ce66881c3ee188a1
