### 2018-10-08 bihealth/sodar-core#3 @mikkonie
mentioned in commit 3b69306930a1db47067af8944e74b5b22815ac39
