### 2018-10-08 bihealth/sodar-core#3 @mikkonie
mentioned in commit 7bfa66e012b08cc56087f4b476502c21f5115868
