### 2018-10-09 bihealth/sodar-core#3 @mikkonie
mentioned in commit ea9ce51f14c213f193dcda3803f9df1f66c7e350
