### 2018-10-09 bihealth/sodar-core#3 @mikkonie
mentioned in commit 62d07106f7ff464e77656c2d7ebd5b720fd0ec26
