### 2018-10-09 bihealth/sodar-core#3 @mikkonie
mentioned in commit 6950ebc3257557f0531de13d817fa845a9e37ab5
