### 2018-10-09 bihealth/sodar-core#3 @mikkonie
mentioned in commit f62c56d6f68791012b185dec961ee2d20fe442b1
