### 2018-10-09 bihealth/sodar-core#3 @mikkonie
mentioned in commit c29cac69dbe5a09304c64070beaabe3d446c9f81
