### 2018-10-09 bihealth/sodar-core#3 @mikkonie
mentioned in commit 124e8b52136fffcb1473553c83d089d1e95a7121
