### 2018-10-10 bihealth/sodar-core#3 @mikkonie
mentioned in commit 8cefddf0bbe0f690cdcedebf926924e11daea627
