### 2018-10-15 bihealth/sodar-core#3 @mikkonie
One potential issue with the current implementation: Empty categories are not synced. So it's not possible to e.g. define rights for a user to own a category and then mirror the category to another site without existing projects.

**TBD:** Is this a problem?
