### 2018-10-18 bihealth/sodar-core#3 @mikkonie
mentioned in commit 079e2615e1fd746b6cb7c60aa9c79d829d79b8a7
