### 2018-10-18 bihealth/sodar-core#3 @mikkonie
An initial implementation of this has been done, with some limitations:

* A site must be set in a SOURCE or TARGET mode. A SOURCE site can define 0-n target sites to provide information to, a TARGET site can define one SOURCE site from which to receive project data.
* There are several modes of project data retrieval, but only two "no access" and "read roles" are currently fully implemented in the retrieving end. These can be implemented and more added as requirements arise.
* Empty categories are not currently synced as in the UI it's only possible to grant "read roles" access to projects. Once that is done, the parent categories are (of course) also synced. However, this doesn't allow us to e.g. define a category with a certain user's ownership, then sync that category to a site where that user can create their own projects at will.

All of these issues are something which can be handled with expanding the current code. The data model and API are very flexible and should not prohibit any of this functionality.
