### 2018-10-18 bihealth/sodar-core#3 @mikkonie
Considering this done. Some bug fixes, additional tests and minor tweaks may still be performed.
