### 2018-10-18 bihealth/sodar-core#3 @mikkonie
closed
