### 2018-11-21 bihealth/sodar-core#6 @mikkonie
mentioned in commit 254cbb12e22b1eeca1b6d2351f327211b8c156f0
