### 2018-11-21 bihealth/sodar-core#6 @mikkonie
This one was partially implemented by the update from @holtgrewe, allowing the display of extra fields in the description.

I think eventually we should still add a display for the entire extra data structure in case there are fields we don't want to put in the description field. Keeping this open for now.
