### 2019-07-07 #1547 @whedon
[PRE REVIEW]: SODAR Core
**Submitting author:** @mikkonie (<a href="http://orcid.org/0000-0002-4180-8810">Mikko Nieminen</a>)
**Repository:** <a href="https://github.com/bihealth/sodar_core" target ="_blank">https://github.com/bihealth/sodar_core</a>
**Version:** v0.6.2
