### 2019-07-07 #1547 @labarba
👋 @mikkonie — your paper is failing to compile because we need the bibliography in a `.bib` file. Please see the section on [https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) in the documentation.
