### 2019-07-09 #1547 @lpantano
Hi @mikkonie , can you add as well a link with the documentation. I don't know if the idea is to be compatible with readthedocs or similar, but to have formatted documentation is important. Thanks!
