### 2019-07-10 #1547 @holtgrewe
Hi @lpantano 👋

(I'm a coauthor of @mikkonie) on this article.

We have a readthedocs documentation linked as "build passing" badge copied from the readthedocs generated markup, would something else be more appropriate?

- https://sodar-core.readthedocs.io/en/latest/?badge=latest

I'll look into the `.bib` file and failing build ASAP.
