### 2019-07-10 #1547 @lpantano
Thanks @holtgrewe , can you add the link to the line `See docs for detailed documentation on use, integration and development.` I didn't see the badget was clickable and I was wondering where the link was. Thank you.
