### 2019-07-10 #1547 @holtgrewe
@lpantano OK adjusted the `README.rst` and split out the bibliography into `.bib`.
