### 2019-07-15 #1547 @holtgrewe
@lpantano Hm, I don't know anyone using Django for life science databases, sadly. Maybe @johanneskoester could recommend someone?
