### 2019-07-19 #1584 @whedon
[REVIEW]: SODAR Core
**Submitting author:** @mikkonie (<a href="http://orcid.org/0000-0002-4180-8810">Mikko Nieminen</a>)
**Repository:** <a href="https://github.com/bihealth/sodar_core" target ="_blank">https://github.com/bihealth/sodar_core</a>
**Version:** v0.8.4
