### 2019-07-22 bihealth/sodar_core#3 @olgabot
Add installation instructions in README and docs
Hello,
Please add installation instructions using an automated package manager such as `pip` or `conda` to the README and documentation. Currently, the docs provide installation instructions only for apps within the project, and the project installation steps are in the "Development" section and require multiple manual steps.

RE https://github.com/openjournals/joss-reviews/issues/1584

Thank you!
Warmest,
Olga
