### 2019-07-22 bihealth/sodar_core#4 @olgabot
Add example workflow or use case
In reading the [JOSS](https://github.com/openjournals/joss-papers/blob/joss.01584/joss.01584/10.21105.joss.01584.pdf) paper and looking over the repo, it seems extremely useful, and I'm having trouble forming in my mind an explicit example workflow using SODAR core. Would it be possible to add an example workflow in either a diagram or prose format? e.g. "SODAR core can be used to upload CSVs from plate readers, manage Illumina samplesheets from these samples, and link them to their sequencing data and results." (I just made this one up). Or something like a ["user story"](http://www.mountaingoatsoftware.com/agile/user-stories) format would be great, too.

RE: https://github.com/openjournals/joss-reviews/issues/1584

Thank you!
Warmest,
Olga
