### 2019-07-22 bihealth/sodar_core#3 @rbatorsky
Hello,
I'm also reviewing and was about to open a related issue so I will comment here instead.  It appears that in order for the installation instructions to be complete, the instructions here https://github.com/bihealth/sodar_django_site should be included more prominently on the main site https://sodar-core.readthedocs.io/en/latest/.

I have successfully installed and connected to the server using these instructions for the example site.  However, it was not straightforward to find.

Thank you,
Rebecca
