### 2019-07-22 bihealth/sodar_core#5 @rbatorsky
Instructions for use within Docker container
Hello,

As part of JOSS review https://github.com/openjournals/joss-reviews/issues/1584 I noticed that the operating system requirements are very restrictive: Ubuntu 16.04.

Since I run a mac, I installed sodar_core running sodar_django_site within a Docker image. Everything works as expected within the container, except I had to change the ip in the container from 127.0.0.1 to 0.0.0.0 because the way that mac runs containers makes it difficult to connect to a specific ip within the container.

 Would the authors be open to including an example using Docker for users that do not run that specific operating system? It would help many people use this interesting piece of software.

Best,
Rebecca
