### 2019-07-22 bihealth/sodar_core#5 @holtgrewe
Dear @rbatorsky,

that's a good catch (answering here as @mikkonie is out of office at the moment).

Actually, I'm 99% certain that it would work out of the box on any Unixoid system (and even Windows wit Linux sub system or the necessary legwork to get the other dependencies working; I managed to get both to work for a SODAR-core based app). However, the developers are all on Ubuntu 16.04 (although deployment goes to CentOS 7) and we might have a hard time to track down problems people have with exotic OS.

Would adding a line "SODAR-Core is developed and tested on Ubuntu 16.04 but most recent versions Linux/Mac/Unix should work (and if you are experienced and try hard you can probably make it work on Windows) although your mileage might vary" convey this sufficiently and friendly enough?

Thanks,
Manuel
