### 2019-07-22 bihealth/sodar_core#4 @holtgrewe
Excellent, we'll add this once @mikkonie is back in office.
