### 2019-07-22 bihealth/sodar_core#3 @holtgrewe
Thanks for pointing this out. I guess this is one of the obvious things for a pair of "fresh" eyes to see while it's hard to find for the actual authors.
