### 2019-07-22 bihealth/sodar_core#4 @rbatorsky
Hello,

Please excuse that I am commenting again, as I hope that my issue is related enough to this one.  An example workflow would be very helpful. Specifically, once I get [sodar_django_site](https://github.com/bihealth/sodar_django_site) up and running and log in, I am not sure what to do!

I have read through the documentation, and am pretty sure that I I create a category, then a project, then invite a user.  However, I'm struggling to find step by step instructions for a new site.  The examples under [Usage](https://sodar-core.readthedocs.io/en/latest/app_projectroles_usage.html) show a site already populated with projects.

Thank you,
Rebecca
