### 2019-07-22 bihealth/sodar_core#5 @rbatorsky
Hi @holtgrewe,
That's good to know. Adding that phase seems reasonable to me.
Best,
Rebecca
