### 2019-07-22 bihealth/sodar_core#4 @holtgrewe
Hi @rbatorsky no, that's an excellent question (as this indicates that the documentation needs improvement, so far we've locally had @mikkonie walk us through things and have not had an external user). However, I'm not 100% positive that I understand what you mean. Maybe the nomenclauture in the documentation could be confusing if you're new to the system.

The word "site" is used in the following contents in the documentation page that you linked to:

- `Django site` -- Your django "web app" that uses SODAR-core (what you created in your derived/modifeid copy of sodar_django_site).
- `their LDAP/AD credentials from a supported site` -- An LDAP/ActiveDirectory server/domain.
- `admin site` -- The django administration UI.
- `site-wide apps` -- A kind of "SODAR Core" app that, in contrast to "project apps" is not bound to a specific project but has "installation-wide scope" of "site" (the "django web app" from above)
- `site info` -- Information about your installation, a site-wide app.
- `When setting up a new site` -- Your installation.
- `Remote Site Access`/`source site`/`target site` -- As explained in the documentation, you can have the installation of a SODAR-based web application marked as "target" that can then sync over projects, roles, and user information from a "source" instance of a SODAR-based web application. One of the very nice features is that this does not have to be the same source code but you can actually have a "central data repository app" that is the source and then "specialized satellite apps" that are connected upstream and sync the projects that they are allowed to. We are using this feature for setting up a central "SODAR" instance with Omics data and meta data (to be published once it's complete but that might take some months) and then, e.g., have a [dashboard hosting web app](https://github.com/bihealth/kiosc) and a [variant filtration web app](https://github.com/bihealth/varfish-server) syndicating information.

[If I'm misrepresenting anything here, I'm certain @mikkonie will correct me when he's back.]

I hope this comment helps. I'm agreeing that the documentation can be improved here, maybe with a more step-by-step workflow. We'll go back to our text editors and think how we could improve this. In the meantime, are you lacking any information to proceed in your review right now?
