### 2019-07-29 bihealth/sodar_core#4 @rbatorsky
Thank you again for your reply. My usage of the word site is the first of the ones you listed: ```Django site```.  I look forward to reading the updated documentation so that I can complete the final steps of the review.
