### 2019-08-09 bihealth/sodar_core#4 @mikkonie
Hi @rbatorsky.  As @holtgrewe said, you raise very good points here. Next week I'll be working on improving the documentation from this perspective.

So far I've identified the following improvements which should help with this issue:

- Add a user story to help describe in how to use SODAR Core in research projects
- Add a general glossary to clarify terms in one place (e.g. the site issue)
- Make sure the site terminology is clearly used within the documentation
- Write out a more high level and user friendly description of the project into the "getting started" section
- Look into adding some helpful graphs to visualize the "big picture", architecture etc.

Please let me know if you would also like to see something else to help in clarifying these issues.
