### 2019-08-09 bihealth/sodar_core#3 @mikkonie
Just to confirm, I will also be working on this. Installing the package via pip will be introduced shortly.
