### 2019-08-12 #1584 @lpantano
Hi @mikkonie, wanted to check if everything is ok, and you are not waiting for us but trying to address the issues opened by the reviewers.

Hi @olgabot,  any chance you could update the checklist above for the items that are ok?

Thank you all!
