### 2019-08-27 bihealth/sodar-core#6 @mikkonie
mentioned in issue sodar-server##639
