### 2019-08-28 bihealth/sodar-core#6 @mikkonie
At least initially, the following display should probably be sufficient:

- Show the actual JSON/dict structure in a modal
- Add button to copy the content into the clipboard
