### 2019-09-01 bihealth/sodar-core#6 @mikkonie
In GitLab by @hendrik.bomhardt on Sep 1, 2019, 15:25

created merge request !83 to address this issue
