### 2019-09-09 bihealth/sodar-core#6 @mikkonie
marked the task **In the timeline list view, add a small link/button for viewing `extra_data` for each `ProjectEvent`, if such data exists** as completed
