### 2019-09-09 bihealth/sodar-core#6 @mikkonie
marked the task **Make the link open a modal** as completed
