### 2019-09-09 bihealth/sodar-core#6 @mikkonie
marked the task **In the modal, view the `extra_data` as nicely formatted raw JSON** as completed
