### 2019-09-09 bihealth/sodar-core#6 @mikkonie
marked the task **For an event, also display related `ProjectEventStatus.extra_data` if such data exists (see below)** as completed
