### 2019-09-16 bihealth/sodar-core#6 @mikkonie
marked the task **Add the possibility to easily copy the JSON into clipboard** as completed
