### 2019-09-16 bihealth/sodar-core#6 @mikkonie
marked the task **The usual test and doc updates** as completed
