### 2019-09-16 bihealth/sodar-core#6 @mikkonie
closed
