### 2019-09-26 bihealth/sodar_core#3 @mikkonie
@olgabot and @rbatorsky: Apologies for the delay in reacting to this. In the currently pushed master branch, installation instructions have been added to both the [README](https://github.com/bihealth/sodar_core/blob/master/README.rst) and the [documentation](https://github.com/bihealth/sodar_core/blob/master/docs/source/getting_started.rst). I hope this simplifies the issue.
