### 2019-09-26 bihealth/sodar_core#4 @mikkonie
@olgabot and @rbatorsky: I've updated the documentation in an attempt to clarify the issues raised here.

- There is a new [Overview section](https://github.com/bihealth/sodar_core/blob/master/docs/source/overview.rst) providing a bit more context and an example use case.
- Also added is the [Glossary](https://github.com/bihealth/sodar_core/blob/master/docs/source/glossary.rst) to clarify some commonly used terms.
- Furthermore, I made multiple general improvements aimed at documenting the package more clearly, most of these can be found in [this commit](https://github.com/bihealth/sodar_core/commit/a638303141d300bfcb1fcb08d22c9d4d00fb7d3e).

I hope this is helpful with the issue at hand, please let me know if I should improve the docs regarding these questions.
