### 2019-09-26 bihealth/sodar_core#5 @mikkonie
@rbatorsky: I added mentions regarding the supported operating system(s) and docker use as discussed above. They can be found in [this commit](https://github.com/bihealth/sodar_core/commit/a638303141d300bfcb1fcb08d22c9d4d00fb7d3e).
