### 2019-10-08 #1584 @mikkonie
Hello,

Apologies for the radio silence. I have actually made changes as suggested by your comments into the master branch of the project and commented on these changes in the issue tickets raised in our project. I missed the updates here and didn't realize I also had to respond in this issue.

Again, sorry about this. Please let me know if you have further comments regarding the issues as they've currently been handled.
