### 2019-10-16 bihealth/sodar_core#6 @rbatorsky
installing via pip - please update sodar_django_site
Hello,

I think the documentation on sodar_django_site regarding how to install sodar_core should be brought up to date. Right now this site has:
```
pip install -e git+https://github.com/bihealth/sodar_core.git@v0.7.0#egg=django-sodar-core
```
And sodar_django_site has:
```
$ git clone git+https://github.com/bihealth/sodar_core.git
$ cd sodar_core
$ pip install virtualenv
$ virtualenv -p python3.6 .venv
$ source .venv/bin/activate
$ utility/install_python_dependencies.sh
```

It's a bit tricky to figure out which are the corresponding lines.

Thanks,
Rebecca
