### 2019-10-16 bihealth/sodar_core#4 @rbatorsky
Hello, I've gone through the documentation and found it helpful.  Thanks.
