### 2019-10-16 #1584 @rbatorsky
Hello,
I've gone through and read the updates. I submitted one issue about installation https://github.com/bihealth/sodar_core/issues/6 but other than that I think I've completed my review.
Best,
Rebecca
