### 2019-10-18 #1584 @lpantano
Thank you @rbatorsky, @olgabot remember to check the boxes when you are finishing each point. Thanks!
