### 2019-10-25 bihealth/sodar_core#6 @mikkonie
Hi @rbatorsky,

Those instructions are actually for different use cases: the first depicting the installation of the sodar-core app package to an existing Django site, the latter describing how to install the entire Django site template using the package. The latter corresponds to the [docs for SODAR Core development installation](https://github.com/bihealth/sodar_core/blob/master/docs/source/dev_sodar_core.rst).

But you are totally right, the instructions are incorrect and refer to the wrong package (sodar_core instead of sodar_django_site). I pushed a commit fixing this and clarifying other installation instructions [here](https://github.com/bihealth/sodar_django_site/commit/dd11779165cd8b44a6dfc4136e1fd0a4fce9941e).

Hopefully this fix is OK for you, please let me know if it should be improved further.
