### 2019-10-28 #1584 @mikkonie
Hello,

FYI, I made changes according to @rbatorsky's new issue and detailed them in the issue comments.

-Mikko
