### 2019-11-02 bihealth/sodar_core#4 @olgabot
Thank you for this! Unfortunately this is not enough for a Django newbie like me. I was able to install the package just fine using the `pip install` command on the [Getting Started page](https://sodar-core.readthedocs.io/en/latest/getting_started.html) and from there I see I need to configure the `projectroles` app but I don't understand what command to run. E.g. I see `django-sodar-core` referenced several times but can't tell if this is a command to run, and it seems like it is not:

```
(sodar_core)
 ♥ 30%  Fri  1 Nov - 18:48  ~ 
  django-sodar-core
zsh: command not found: django-sodar-core
```

Would it be possible to add a full use case with all the commands and their outputs, e.g. for the servers you have already set up?

### Gory installation details

```
(base)
 ✘  ♥ 39%  Fri  1 Nov - 18:33  ~ 
  conda create -n sodar_core python=3.7
Collecting package metadata (current_repodata.json): done
Solving environment: done

## Package Plan ##

  environment location: /anaconda3/envs/sodar_core

  added / updated specs:
    - python=3.7

The following packages will be downloaded:

    package                    |            build
    ---------------------------|-----------------
    libcxx-9.0.0               |       h89e68fa_1         1.0 MB  conda-forge
    pip-19.3.1                 |           py37_0         1.9 MB  conda-forge
    setuptools-41.6.0          |           py37_1         650 KB  conda-forge
    ------------------------------------------------------------
                                           Total:         3.5 MB

The following NEW packages will be INSTALLED:

  bzip2              conda-forge/osx-64::bzip2-1.0.8-h01d97ff_1
  ca-certificates    conda-forge/osx-64::ca-certificates-2019.9.11-hecc5488_0
  certifi            conda-forge/osx-64::certifi-2019.9.11-py37_0
  libcxx             conda-forge/osx-64::libcxx-9.0.0-h89e68fa_1
  libffi             bioconda/osx-64::libffi-3.2.1-1
  ncurses            conda-forge/osx-64::ncurses-6.1-h0a44026_1002
  openssl            conda-forge/osx-64::openssl-1.1.1c-h01d97ff_0
  pip                conda-forge/osx-64::pip-19.3.1-py37_0
  python             conda-forge/osx-64::python-3.7.3-h93065d6_1
  readline           conda-forge/osx-64::readline-8.0-hcfe32e1_0
  setuptools         conda-forge/osx-64::setuptools-41.6.0-py37_1
  sqlite             conda-forge/osx-64::sqlite-3.30.1-h93121df_0
  tk                 conda-forge/osx-64::tk-8.6.9-h2573ce8_1003
  wheel              conda-forge/osx-64::wheel-0.33.6-py37_0
  xz                 conda-forge/osx-64::xz-5.2.4-h1de35cc_1001
  zlib               conda-forge/osx-64::zlib-1.2.11-h0b31af3_1006

Downloading and Extracting Packages
setuptools-41.6.0    | 650 KB    | ##################################################################################################################################################################### | 100%
pip-19.3.1           | 1.9 MB    | ##################################################################################################################################################################### | 100%
libcxx-9.0.0         | 1.0 MB    | ##################################################################################################################################################################### | 100%
Preparing transaction: done
Verifying transaction: done
Executing transaction: done
#
# To activate this environment, use
#
#     $ conda activate sodar_core
#
# To deactivate an active environment, use
#
#     $ conda deactivate

(base)
 ♥ 36%  Fri  1 Nov - 18:35  ~ 
  conda activate sodar_core
(sodar_core)
 ♥ 36%  Fri  1 Nov - 18:40  ~ 
  pip install -e git+https://github.com/bihealth/sodar_core.git@v0.7.0#egg=django-sodar-core

Obtaining django-sodar-core from git+https://github.com/bihealth/sodar_core.git@v0.7.0#egg=django-sodar-core
  Cloning https://github.com/bihealth/sodar_core.git (to revision v0.7.0) to ./src/django-sodar-core
  Running command git clone -q https://github.com/bihealth/sodar_core.git /Users/olgabot/src/django-sodar-core
Collecting django-autocomplete-light==3.2.10
  Downloading https://files.pythonhosted.org/packages/ee/2c/8830276bfdb01713486585640125e0b6d0f2b4ee2ece96689a846c3d77cc/django-autocomplete-light-3.2.10.tar.gz (428kB)
     |████████████████████████████████| 430kB 72kB/s
Collecting django-db-file-storage==0.5.3
  Downloading https://files.pythonhosted.org/packages/a0/99/1d2bcb6255c7552e73637097dfefb99c94f22800ae254a02e32e77905068/django-db-file-storage-0.5.3.tar.gz
Collecting pytz>=2019.3
  Downloading https://files.pythonhosted.org/packages/e7/f9/f0b53f88060247251bf481fa6ea62cd0d25bf1b11a87888e53ce5b7c8ad2/pytz-2019.3-py2.py3-none-any.whl (509kB)
     |████████████████████████████████| 512kB 111kB/s
Collecting django-model-utils<3.3,>=3.2.0
  Downloading https://files.pythonhosted.org/packages/7e/ed/c8d4654bce5859766c989ca0cc3dc439d3abd00f5b1c6ef4d80b0a5f2871/django_model_utils-3.2.0-py2.py3-none-any.whl
Requirement already satisfied: wheel==0.33.6 in /anaconda3/envs/sodar_core/lib/python3.7/site-packages (from django-sodar-core) (0.33.6)
Collecting django<2.0,>=1.11.25
  Downloading https://files.pythonhosted.org/packages/3c/ed/06a81a65fa00f766f2dbda94d09e946aa65c23e6d7ca3532984627a6c75a/Django-1.11.25-py2.py3-none-any.whl (6.9MB)
     |████████████████████████████████| 7.0MB 62kB/s
Collecting django-pagedown<1.1,>=1.0.6
  Downloading https://files.pythonhosted.org/packages/78/72/a1974529247f3e584d8653d8ff6707d0e807f62bbe349d253b4b9188989b/django-pagedown-1.0.6.tar.gz (79kB)
     |████████████████████████████████| 81kB 22kB/s
Collecting Pillow<6.3,>=6.2.0
  Downloading https://files.pythonhosted.org/packages/85/28/2c72ba965b52884a0bd71e419761fc162763dc2e5d9bec2f3b1949f7272a/Pillow-6.2.1-cp37-cp37m-macosx_10_6_intel.whl (3.9MB)
     |████████████████████████████████| 3.9MB 405kB/s
Collecting argon2-cffi<19.2,>=19.1.0
  Downloading https://files.pythonhosted.org/packages/a0/b2/9e5cd048e4a6df28a7a01380e81e9066350757d3e031feda241d231a7ea2/argon2_cffi-19.1.0-cp34-abi3-macosx_10_6_intel.whl (59kB)
     |████████████████████████████████| 61kB 523kB/s
Collecting django-rest-knox<4.2,>=4.1.0
  Downloading https://files.pythonhosted.org/packages/2b/99/dd6e9c75252f1cdea28d275f27ba7641c14226d37016ecf873eda03da105/django_rest_knox-4.1.0-py3-none-any.whl
Collecting markdown<3.2,>=3.1.1
  Downloading https://files.pythonhosted.org/packages/c0/4e/fd492e91abdc2d2fcb70ef453064d980688762079397f779758e055f6575/Markdown-3.1.1-py2.py3-none-any.whl (87kB)
     |████████████████████████████████| 92kB 954kB/s
Collecting awesome-slugify<1.7,>=1.6.5
  Downloading https://files.pythonhosted.org/packages/34/39/79ef4e640c3651b40de7812f5fcd04698abf14de4f57a81e12b6c753d168/awesome-slugify-1.6.5.tar.gz
Collecting rules<2.2,>=2.1
  Downloading https://files.pythonhosted.org/packages/8a/5d/f35209aeb1e34e6cf0ad113930829618864a8a455ed2a9c63c696d822f97/rules-2.1.tar.gz (52kB)
     |████████████████████████████████| 61kB 597kB/s
Collecting djangorestframework<3.11,>=3.10.3
  Downloading https://files.pythonhosted.org/packages/33/8e/87a4e0025e3c4736c1dc728905b1b06a94968ce08de15304417acb40e374/djangorestframework-3.10.3-py3-none-any.whl (909kB)
     |████████████████████████████████| 911kB 867kB/s
Collecting django-docs<0.4,>=0.3.1
  Downloading https://files.pythonhosted.org/packages/60/a3/e5911935c9476850d9c828cdcd5f36c3e3c4b09f6ed087d361e0dfdda238/django_docs-0.3.1-py3-none-any.whl
Collecting versioneer>=0.18
  Downloading https://files.pythonhosted.org/packages/95/b5/8bcf39663abc1fda6a2af6704062a44be13d3bc1ceca562a9c020fae8f36/versioneer-0.18-py2.py3-none-any.whl
Collecting psycopg2-binary<2.9,>=2.8.3
  Downloading https://files.pythonhosted.org/packages/13/c3/050b8da85b02886f2c649da42cf195a4dc8cc5770a5524cd5b34c36c03a7/psycopg2_binary-2.8.4-cp37-cp37m-macosx_10_6_intel.macosx_10_9_intel.macosx_10_9_x86_64.macosx_10_10_intel.macosx_10_10_x86_64.whl (1.5MB)
     |████████████████████████████████| 1.6MB 22kB/s
Collecting celery<4.2
  Downloading https://files.pythonhosted.org/packages/99/fa/4049b26bfe71992ecf979acd39b87e55b493608613054089d975418015b7/celery-4.1.1-py2.py3-none-any.whl (394kB)
     |████████████████████████████████| 399kB 518kB/s
Collecting django-markupfield<1.6,>=1.5.1
  Downloading https://files.pythonhosted.org/packages/e9/d9/47cbfc67c73a94a905add1b294c9fa6e48b53ad44b100133f12933eb03b4/django_markupfield-1.5.1-py2.py3-none-any.whl
Collecting mistune<0.9,>=0.8.4
  Using cached https://files.pythonhosted.org/packages/09/ec/4b43dae793655b7d8a25f76119624350b4d65eb663459eb9603d7f1f0345/mistune-0.8.4-py2.py3-none-any.whl
Collecting django-crispy-forms<1.8,>=1.7.2
  Downloading https://files.pythonhosted.org/packages/9a/05/6bad05742d185ec2fabfa4deab05cafde286eb3f383fba24b3674340aca2/django_crispy_forms-1.7.2-py2.py3-none-any.whl (105kB)
     |████████████████████████████████| 112kB 840kB/s
Collecting Sphinx<2.3,>=2.2.0
  Downloading https://files.pythonhosted.org/packages/1d/58/a547b02946af1cf337c701cafcf3267065c1c83d6e701f540961f4907b5d/Sphinx-2.2.1-py3-none-any.whl (3.3MB)
     |████████████████████████████████| 3.3MB 288kB/s
Collecting sphinx-rtd-theme<0.5,>=0.4.3
  Downloading https://files.pythonhosted.org/packages/60/b4/4df37087a1d36755e3a3bfd2a30263f358d2dea21938240fa02313d45f51/sphinx_rtd_theme-0.4.3-py2.py3-none-any.whl (6.4MB)
     |████████████████████████████████| 6.4MB 1.2MB/s
Collecting django-environ<0.5,>=0.4.5
  Downloading https://files.pythonhosted.org/packages/9f/32/76295a1a5d00bf556c495216581c6997e7fa5f533b2229e0a9d6cbaa95ae/django_environ-0.4.5-py2.py3-none-any.whl
Collecting six
  Using cached https://files.pythonhosted.org/packages/73/fb/00a976f728d0d1fecfe898238ce23f502a721c0ac0ecfedb80e0d88c64e9/six-1.12.0-py2.py3-none-any.whl
Collecting cffi>=1.0.0
  Downloading https://files.pythonhosted.org/packages/41/bd/61e98c053fd54317decdc314279a0917e1c3807186f0ee7ea854cd09a645/cffi-1.13.1-cp37-cp37m-macosx_10_6_intel.whl (258kB)
     |████████████████████████████████| 266kB 609kB/s
Collecting cryptography
  Downloading https://files.pythonhosted.org/packages/6b/4a/ce93178469d4460d6b3a5e648fc1a2f426030f3d30a12d7ed4df73d044de/cryptography-2.8-cp34-abi3-macosx_10_6_intel.whl (1.6MB)
     |████████████████████████████████| 1.6MB 1.0MB/s
Requirement already satisfied: setuptools>=36 in /anaconda3/envs/sodar_core/lib/python3.7/site-packages (from markdown<3.2,>=3.1.1->django-sodar-core) (41.6.0.post20191101)
Collecting regex
  Downloading https://files.pythonhosted.org/packages/fc/1d/13cc7d174cd2d05808abac3f5fb37433e30c4cd93b152d2a9c09c926d7e8/regex-2019.11.1.tar.gz (669kB)
     |████████████████████████████████| 675kB 1.1MB/s
Collecting Unidecode<0.05,>=0.04.14
  Downloading https://files.pythonhosted.org/packages/01/a1/9d7f3138ee3d79a1ab865a2cb38200ca778d85121db19fe264c76c981184/Unidecode-0.04.21-py2.py3-none-any.whl (228kB)
     |████████████████████████████████| 235kB 709kB/s
Collecting kombu<5.0,>=4.2.0
  Downloading https://files.pythonhosted.org/packages/42/16/f82ef8692fbd30ea7711e92e1775bb237888f4cef3a0d24dcf33a1968834/kombu-4.6.5-py2.py3-none-any.whl (182kB)
     |████████████████████████████████| 184kB 852kB/s
Collecting billiard<3.6.0,>=3.5.0.2
  Downloading https://files.pythonhosted.org/packages/8b/b7/c2fe04f2522bb02d044347734eeda3ff5c7a632fa7d0401530a371ba73db/billiard-3.5.0.5.tar.gz (150kB)
     |████████████████████████████████| 153kB 1.1MB/s
Collecting Jinja2>=2.3
  Downloading https://files.pythonhosted.org/packages/65/e0/eb35e762802015cab1ccee04e8a277b03f1d8e53da3ec3106882ec42558b/Jinja2-2.10.3-py2.py3-none-any.whl (125kB)
     |████████████████████████████████| 133kB 1.5MB/s
Collecting requests>=2.5.0
  Using cached https://files.pythonhosted.org/packages/51/bd/23c926cd341ea6b7dd0b2a00aba99ae0f828be89d72b2190f27c11d4b7fb/requests-2.22.0-py2.py3-none-any.whl
Collecting sphinxcontrib-devhelp
  Using cached https://files.pythonhosted.org/packages/b0/a3/fea98741f0b2f2902fbf6c35c8e91b22cd0dd13387291e81d457f9a93066/sphinxcontrib_devhelp-1.0.1-py2.py3-none-any.whl
Collecting imagesize
  Using cached https://files.pythonhosted.org/packages/fc/b6/aef66b4c52a6ad6ac18cf6ebc5731ed06d8c9ae4d3b2d9951f261150be67/imagesize-1.1.0-py2.py3-none-any.whl
Collecting babel!=2.0,>=1.3
  Downloading https://files.pythonhosted.org/packages/2c/60/f2af68eb046c5de5b1fe6dd4743bf42c074f7141fe7b2737d3061533b093/Babel-2.7.0-py2.py3-none-any.whl (8.4MB)
     |████████████████████████████████| 8.4MB 3.1MB/s
Collecting sphinxcontrib-htmlhelp
  Using cached https://files.pythonhosted.org/packages/e4/35/80a67cc493f4a8a9634ab203a77aaa1b84d79ccb1c02eca72cb084d2c7f7/sphinxcontrib_htmlhelp-1.0.2-py2.py3-none-any.whl
Collecting sphinxcontrib-serializinghtml
  Using cached https://files.pythonhosted.org/packages/57/b3/3648e48fa5682e61e9839d62de4e23af1795ceb738d68d73bd974257a95c/sphinxcontrib_serializinghtml-1.1.3-py2.py3-none-any.whl
Collecting alabaster<0.8,>=0.7
  Using cached https://files.pythonhosted.org/packages/10/ad/00b090d23a222943eb0eda509720a404f531a439e803f6538f35136cae9e/alabaster-0.7.12-py2.py3-none-any.whl
Collecting Pygments>=2.0
  Using cached https://files.pythonhosted.org/packages/5c/73/1dfa428150e3ccb0fa3e68db406e5be48698f2a979ccbcec795f28f44048/Pygments-2.4.2-py2.py3-none-any.whl
Collecting snowballstemmer>=1.1
  Downloading https://files.pythonhosted.org/packages/7d/4b/cdf1113a0e88b641893b814e9c36f69a6fda28cd88b62c7f0d858cde3166/snowballstemmer-2.0.0-py2.py3-none-any.whl (97kB)
     |████████████████████████████████| 102kB 2.3MB/s
Collecting docutils>=0.12
  Using cached https://files.pythonhosted.org/packages/22/cd/a6aa959dca619918ccb55023b4cb151949c64d4d5d55b3f4ffd7eee0c6e8/docutils-0.15.2-py3-none-any.whl
Collecting sphinxcontrib-jsmath
  Using cached https://files.pythonhosted.org/packages/c2/42/4c8646762ee83602e3fb3fbe774c2fac12f317deb0b5dbeeedd2d3ba4b77/sphinxcontrib_jsmath-1.0.1-py2.py3-none-any.whl
Collecting sphinxcontrib-applehelp
  Using cached https://files.pythonhosted.org/packages/13/9a/4428b3114d654cb1cd34d90d5e6fab938d5436f94a571155187ea1dd78b4/sphinxcontrib_applehelp-1.0.1-py2.py3-none-any.whl
Collecting packaging
  Using cached https://files.pythonhosted.org/packages/cf/94/9672c2d4b126e74c4496c6b3c58a8b51d6419267be9e70660ba23374c875/packaging-19.2-py2.py3-none-any.whl
Collecting sphinxcontrib-qthelp
  Using cached https://files.pythonhosted.org/packages/ce/5b/4747c3ba98b3a3e21a66faa183d8f79b9ded70e74212a7988d236a6eb78a/sphinxcontrib_qthelp-1.0.2-py2.py3-none-any.whl
Processing ./Library/Caches/pip/wheels/f2/9a/90/de94f8556265ddc9d9c8b271b0f63e57b26fb1d67a45564511/pycparser-2.19-py2.py3-none-any.whl
Collecting importlib-metadata>=0.18
  Using cached https://files.pythonhosted.org/packages/f6/d2/40b3fa882147719744e6aa50ac39cf7a22a913cbcba86a0371176c425a3b/importlib_metadata-0.23-py2.py3-none-any.whl
Collecting amqp==2.5.1
  Downloading https://files.pythonhosted.org/packages/a1/ef/a62ff9dde89e6b75976a746aa5b828ccb3c839ac6fcdc283edd88d7cf90b/amqp-2.5.1-py2.py3-none-any.whl (49kB)
     |████████████████████████████████| 51kB 1.9MB/s
Collecting MarkupSafe>=0.23
  Using cached https://files.pythonhosted.org/packages/ce/c6/f000f1af136ef74e4a95e33785921c73595c5390403f102e9b231b065b7a/MarkupSafe-1.1.1-cp37-cp37m-macosx_10_6_intel.whl
Collecting urllib3!=1.25.0,!=1.25.1,<1.26,>=1.21.1
  Using cached https://files.pythonhosted.org/packages/e0/da/55f51ea951e1b7c63a579c09dd7db825bb730ec1fe9c0180fc77bfb31448/urllib3-1.25.6-py2.py3-none-any.whl
Collecting idna<2.9,>=2.5
  Using cached https://files.pythonhosted.org/packages/14/2c/cd551d81dbe15200be1cf41cd03869a46fe7226e7450af7a6545bfc474c9/idna-2.8-py2.py3-none-any.whl
Collecting chardet<3.1.0,>=3.0.2
  Using cached https://files.pythonhosted.org/packages/bc/a9/01ffebfb562e4274b6487b4bb1ddec7ca55ec7510b22e4c51f14098443b8/chardet-3.0.4-py2.py3-none-any.whl
Requirement already satisfied: certifi>=2017.4.17 in /anaconda3/envs/sodar_core/lib/python3.7/site-packages (from requests>=2.5.0->Sphinx<2.3,>=2.2.0->django-sodar-core) (2019.9.11)
Collecting pyparsing>=2.0.2
  Using cached https://files.pythonhosted.org/packages/11/fa/0160cd525c62d7abd076a070ff02b2b94de589f1a9789774f17d7c54058e/pyparsing-2.4.2-py2.py3-none-any.whl
Collecting zipp>=0.5
  Using cached https://files.pythonhosted.org/packages/74/3d/1ee25a26411ba0401b43c6376d2316a71addcc72ef8690b101b4ea56d76a/zipp-0.6.0-py2.py3-none-any.whl
Collecting vine<5.0.0a1,>=1.1.3
  Downloading https://files.pythonhosted.org/packages/7f/60/82c03047396126c8331ceb64da1dc52d4f1317209f32e8fe286d0c07365a/vine-1.3.0-py2.py3-none-any.whl
Collecting more-itertools
  Using cached https://files.pythonhosted.org/packages/45/dc/3241eef99eb45f1def35cf93af35d1cf9ef4c0991792583b8f33ea41b092/more_itertools-7.2.0-py3-none-any.whl
Building wheels for collected packages: django-autocomplete-light, django-db-file-storage, django-pagedown, awesome-slugify, rules, regex, billiard
  Building wheel for django-autocomplete-light (setup.py) ... done
  Created wheel for django-autocomplete-light: filename=django_autocomplete_light-3.2.10-cp37-none-any.whl size=514922 sha256=4a0d3b56bb92bf4c718d6e7ae937ba5451ffef027e60aabe3aa0e56f56a714c0
  Stored in directory: /Users/olgabot/Library/Caches/pip/wheels/15/7a/ec/435379d140de664ba9f723dab0abb09c8c0da62ef09fc48526
  Building wheel for django-db-file-storage (setup.py) ... done
  Created wheel for django-db-file-storage: filename=django_db_file_storage-0.5.3-cp37-none-any.whl size=9060 sha256=95a1703bac70849222f858d4613b3f61a6843bb4500a1744f148bfc71fa306fd
  Stored in directory: /Users/olgabot/Library/Caches/pip/wheels/63/8e/ce/9be5f41d1bce93a68373f7ccc941f98060022735fc0319a4c1
  Building wheel for django-pagedown (setup.py) ... done
  Created wheel for django-pagedown: filename=django_pagedown-1.0.6-cp37-none-any.whl size=81656 sha256=fc5a08f51637e8c621f1e2ce9888c5cf69d9f8a0e33c51c70dc596ddde91ac41
  Stored in directory: /Users/olgabot/Library/Caches/pip/wheels/66/d6/42/129813de197dff6dbd32ef142a31763bc006edf27e693c00f1
  Building wheel for awesome-slugify (setup.py) ... done
  Created wheel for awesome-slugify: filename=awesome_slugify-1.6.5-cp37-none-any.whl size=8351 sha256=0ba855ece058096a1b0903d855804b18ace02836b761cd35ebdac54bb525663f
  Stored in directory: /Users/olgabot/Library/Caches/pip/wheels/e9/d9/66/bde66382496710218c0df31b5f3a72bd8079bcd275fff61b29
  Building wheel for rules (setup.py) ... done
  Created wheel for rules: filename=rules-2.1-py2.py3-none-any.whl size=25776 sha256=2e19feb177f4bc48d93a12601cda2c9a8b88677c2cb19ec443526915e7b571fb
  Stored in directory: /Users/olgabot/Library/Caches/pip/wheels/c5/9f/fc/cbf9eff602ec470e1443d68957f9488fa30e30aa5c1a12736b
  Building wheel for regex (setup.py) ... done
  Created wheel for regex: filename=regex-2019.11.1-cp37-cp37m-macosx_10_9_x86_64.whl size=291847 sha256=4a0959973ffa22b5fceb765f74950820695d4656625456083c2c9f8b4a609fa0
  Stored in directory: /Users/olgabot/Library/Caches/pip/wheels/5c/c6/c1/0bc8d16ea38c44536a82dd1bec665996e5af37489fa88826b6
  Building wheel for billiard (setup.py) ... done
  Created wheel for billiard: filename=billiard-3.5.0.5-cp37-none-any.whl size=87882 sha256=6ee1348785c8108339ea063f301ad9f865eecb51d0c3b36c32347225d22d310d
  Stored in directory: /Users/olgabot/Library/Caches/pip/wheels/b8/72/0e/39ecdedc4cfc45b693a623732e40dbd4cff5ea5e11775ee591
Successfully built django-autocomplete-light django-db-file-storage django-pagedown awesome-slugify rules regex billiard
[…]
