### 2019-11-02 bihealth/sodar_core#8 @olgabot
Add community guidelines for getting help and contributing
From this review: https://github.com/openjournals/joss-reviews/issues/1584, I wasn't able to find a community guideline or code of conduct or equivalent for getting help, contributing code, and interacting with one another. Could you let me know if I missed it?

In case you are looking for an example code of conduct, here's a great place to start: https://www.contributor-covenant.org/
