### 2019-11-02 #1584 @olgabot
Took a look at this, and have some questions/concerns:

- The release version is now [0.7.0](https://github.com/bihealth/sodar_core/releases/tag/v0.7.0), but this issue refers to 0.6.2 -- should this get updated in the issue?
- I've never used Django before, and I can't seem to get my mind around what this package does and how to use it, and thus cannot evaluate the functionality or performance. Continued the discussion here: https://github.com/bihealth/sodar_core/issues/4
- "Example usage: Do the authors include examples of how to use the software (**ideally to solve real-world analysis problems**)."  (emphasis mine) -- It's not clear to me how this solves analysis problems. So far, I can tell that there is a web server but I don't understand how to set it up or how it can help solve analysis problems
