### 2019-11-19 #1584 @kthyng
hi  @mikkonie! Have you had a chance to address comments from @olgabot?
