### 2019-12-09 #1584 @mikkonie
@olgabot and @danielskatz, I apologize for the delay in reacting to this.

Regarding the version number: there have indeed been some updates and I expect to tag version 0.7.1 before the holidays. If this does not cause issues from the review side, it sounds logical to update the version. Please let me know if something needs to be clarified regarding the new releases.

As for the usage of the package, I will try to elaborate on the context of the package more in the aforementioned discussion. Hopefully it helps.

On the real world analysis use case issue: there are publications in works which detail real life scientific use cases for this package. Sadly they have not been published yet, so we can't simply refer to them. One option we discussed internally is e.g. providing an example demo application which displays a scientific use case and how the package enables it. Would this sound helpful?
