### 2019-12-09 bihealth/sodar_core#4 @mikkonie
@olgabot: I'm sorry the context of the package remains unclear. It is indeed a complex issue for someone not used to working with the Django framework.

In the Django ecosystem, it is commonplace to have packages which provide additions to the base web framework, either in form of applications, middleware or both. `django-sodar-core` refers to the SODAR Core package, which is installed on top of a Django web site project in the aforementioned manner. It is a set of enhancements over the base Django web framework aimed at developing web based systems for scientific research projects.

Hence the package is not usable by itself, but is expected to be installed to an existing Django web site project. An example of a minimal Django project with the web framework and the django-sodar-core package installed can be seen in the [sodar_django_site](https://github.com/bihealth/sodar_django_site) project. On top of this, the actual research project applications are developed.

As I mentioned [here](https://github.com/openjournals/joss-reviews/issues/1584#issuecomment-563317931), we have several systems using the package for actual scientific apps. Those would provide a view to the entire use case for using this package, but sadly publications on them are still pending.

One possibility would be to develop a demo application, which represents e.g. an analysis process and how it would be implemented on SODAR Core. In case this would sound like a good approach, anything more specific you would be looking for in such an application?
