### 2019-12-09 bihealth/sodar_core#8 @mikkonie
Indeed, these have not been defined. Will look into it, thank you for noting.
