### 2019-12-18 #1584 @mikkonie
As a heads up, I will be on vacation and mostly unavailable until 6.1. After that I'm happy to provide any more help and updates required to further the review. Also, in January it would be possible to use a published use case in the preprint stage in order to help illustrate use cases for this project.
