### 2020-01-09 #1584 @lpantano
Hi @mikkonie, just wanted to check if you still need to work more on the paper before I ping the reviewers again. Just let us know when everything is ready, for a another round. Thanks! and happy new year!.
