### 2020-02-03 bihealth/sodar_core#6 @holtgrewe
@rbatorsky thank you for raising this point. We have restructured the documentation a bit and added a "for the impatient" section here:

- https://sodar-core.readthedocs.io/en/latest/for_the_impatient.html#for-the-impatient

We hope that this step-by-step walkthrough helps people to jump into the action quicker. We'd be happy to address specific issues and questions.

Please re-open this if you feel we have not addressed the issues sufficiently.
