### 2020-02-03 bihealth/sodar_core#4 @holtgrewe
Dear @olgabot and @rbatorsky, thanks for raising these points. As pointed out in #6, we now provide a section that gives the precise commands to setup a sodar-core site based on the template that we provide:

- https://sodar-core.readthedocs.io/en/latest/for_the_impatient.html#for-the-impatient

Please have a look at this whether it addresses all of your questions and points. Please re-open this issue if you feel that we left a point unanswered and this could use some more work.
