### 2020-02-03 bihealth/sodar_core#3 @holtgrewe
@olgabot @rbatorsky We added a link to the quickstart in the README.

- https://github.com/bihealth/sodar_core#quickstart

I'm closing this ticket for now, please re-open in the case that you think we should address this in a different way.
