### 2020-02-03 bihealth/sodar_core#8 @holtgrewe
@olgabot thank you, this is a very good point.

We have now added contribution guidelines and defined a code of conduct

- https://sodar-core.readthedocs.io/en/latest/contributing.html
- https://sodar-core.readthedocs.io/en/latest/code_of_conduct.html
