### 2020-02-19 #1584 @lpantano
@rbatorsky, @olgabot, can you take a look at this and see whether the changes are good to continue.
