### 2020-02-20 bihealth/sodar_core#11 @olgabot
More questions + Add screenshots / visual explanation of what the tool is
Hello,
I am still having trouble understand what this package is and what it does. Some of the comments from https://github.com/bihealth/sodar_core/issues/4 were addressed, but I still don't understand what the app does. The `paper.md` could use some revision to reduce jargon and clarify the language for the non-Django user, to motivate them to invest time in using the app. Currently, as someone who has never written Django code, I don't know what problem this tool solves and don't know how it could improve my workflow.

Some questions:

1. What is some expected "content?"
2. Are workflows/pipelines part of the content?
3. What are different `projectroles` that you envision?
4. Can you show some example input/output of these tools?
5. Could you add "user stories" that guide the reader to understand exactly the kinds of use cases you envision?

I think that some of the questions could also be answered by adding screenshots of the app in action, to show potential use cases, or with a visual diagram showing inputs/outputs.

Finally, do you have an end-to-end test that runs some overarching command? I was only able to find unit tests.

Thank you!
Warmest,
Olga
