### 2020-02-20 #1584 @olgabot
I'm still having trouble understanding what this tool does, and added a issue: https://github.com/bihealth/sodar_core/issues/11
