### 2020-02-20 bihealth/sodar_core#11 @holtgrewe
Hi,

I would propose to clarify some things here in the issues and then I go forward to update the documentation accordingly.
Do you think that would be good?

First of all, sodar-core is a framework/library and not a tool. You use the library and contained software components in the same way as you would use libraries for Django to write your own web apps. The first sentence in the README, the documentation says "SODAR Core is a framework for Django web application development" while there is the following sentence in the third paragraph of the paper.md: "In order to prevent duplication of work, we have extracted the commonly useful components into SODAR Core, a Python framework on top of Django."

It provides a set of components (the Django terminology is "apps" for them) including projectroles.

Does this help with the issue you raised regarding scope the first paragraph of your question?

Regarding your questions:

1. We provide links to three different apps in the end of the [For the Impatient](https://sodar-core.readthedocs.io/en/latest/for_the_impatient.html#going-on-from-here) section.
   That is, VarFish where the content is rare disease genetics cases with germline variants, DigestiFlow, where the content is information for Illumina flow cells, and Kiosc where the content is Docker containers running web dashboards.
   Finally, the filesfolders app provides a use case where people can upload small files, annotate them, and oragnize them in folders.
   We point to filesfolders in the [Summary](https://sodar-core.readthedocs.io/en/latest/for_the_impatient.html#summary) section of the For the Impatient section.
2. There is no workflow or pipeline shipping with sodar-core.
   Users can refer to the three applications VarFish, DigestiFlow, and Kiosc for larger scale examples while the filesfolders app is a nice, small example.
3. I don't understand the question here.
   The projectrole app provides projects and giving users access to the projects with roles.
   From the paper.md: "All SODAR Core features are grouped around projects that can be arranged in a tree structure.
   Users can be assigned to projects with certain roles that can be used for controlling authorization.
   This feature is provided in the `projectroles` app."
4. The For the Impatient section shows with screenshots how an example project looks like.
   Additionall, you can check out https://varfish-demo.bihealth.org/login/?next=/ to see how the VarFish web app looks like that is based on sodar-core and you will see the similarity to the screenshot and how that it extended.
5. What exactly do you mean here?
   A possible existing user story could be VarFish, we could write something along the lines of:
   "The overall aim was to develop a web application that allows biomedical researchers to analyze their data.
   They want organize their patients/cases into projects and only give access to certain users.
   For this, the end user would create a sodar-core project for each study and assign the "contributor" role to the data analysts.
   The developer would then create a new Django app "variants" that integrates with SODAR-core (as described in the documentation).
   This app would then provide the model classes interfacing with the database where the case and variant information is stored.
   The developer would further implement an interface that allows for the data analysis (e.g., different search parameters) and display the results back to the user.
   The end user would then see their cases in their projects and would use this interface to analyze the data."
   Is this something that would help the understanding?

Finally, as this is a library with software components, these are tested. There are no integration or end-to-end tests but that is very uncommon for django libraries (I cannot not think of even one).

Please let me know if any of this would be helpful and in which area you think this would need improvement?

Best wishes,
Manuel
