### 2020-03-04 bihealth/sodar_core#11 @holtgrewe
Dear Olga, please excuse the delay. Because of another big revision we have to postpone continuing with this review. I'll ping you here once we are ready to continue with this and I will come forward with proposals for our updates.
