### 2020-03-04 #1584 @holtgrewe
Hi I also wanted to report that it will take us 4 weeks to get back to this review because of another big revision in our group. We'll continue in the sodar_core ticket tracker with @olgabot
