### 2020-04-27 bihealth/sodar-core#8 @mikkonie
mentioned in issue sodar-server##846
