### 2020-05-20 #1584 @arfon
:wave: @holtgrewe @mikkonie - JOSS [is now open again](https://blog.joss.theoj.org/2020/05/reopening-joss) for submissions and we're actively trying to wrap up existing reviews. Do you think you might be able to address the reviewer feedback soon?
