### 2020-10-26 bihealth/sodar_core#11 @holtgrewe
Dear @olgabot,

I'm picking this up (also in reference to @arfon in his [comment in the ticket](https://github.com/openjournals/joss-reviews/issues/1584#issuecomment-716119568)). I am sorry for the long delay.

Connected to this is PR #14 that addresses these issues in the documentation.

> Thank you, this is helpful. Are there any parallels you could draw, e.g. SODAR Core is a framework for Django like X is a framework for Y?

I added the following to README and manual front page

*In that it is similar to the CMS or ecommerce frameworks that you can find `Awesome Django List <https://github.com/wsvincent/awesome-django#content-management-systems>`__ but you will find the components/libraries provided in SODAR Core are more generic and in this reflecting the broader range of applications that we target.*

would that help?

> If you haven't already, I recommend adding your package to bioconda (#13) as I had trouble installing it in my own environment.

As explained in the reply to #13 we would like to keep the installation instructions agnostic to the used package manager. As mentioned in the [Sodar Core Documentation's front page](https://sodar-core.readthedocs.io/en/latest/#what-s-not-inside-this-documentation) we don't cover basic Python or Django development. In this, SODAR Core targets more advanced Python users, similar to the Django documentation not teaching Python.

Your issue in #13 indicates that you are missing the host dependency of postgres development.
I addressed this in my update to the documentation.
However, I would consider this out of scope as SODAR Core targets people with intermediate experience in writing Django applications and such developers will have solved this kind of problem with elementary libraries earlier in their learning curve of Django and Postgres.

> My issue here is that the examples have been left as an "exercise to the reader," whereas if you hope to gain the most number of users, having explicit examples on the homepage, or in a very clear place like an "example use cases" page, that will help newcomers to the project quickly understand what the project is, what problems it solves, and can more easily decide whether it is useful for them.

Following your suggestion have added a "See It In Action" section to both README and the documentation.
I also added a section "user stories" that gives a very birds-eye view on how we used SODAR Core in our applications.

> Can you give some examples, maybe in the user stories, about how projectroles are used?

OK, adjusted the documentation and paper.md accordingly with user stories secion.

> This definition is circular - the words "project" and "role" are in both the term and the definition. Can you define it in other words?

A "project" is a container for other data. A category is a special project in that it can contain other projects but no other kind of data. A "role" is a "name token" that can be assigned to users in the context of a given project. Users thus have the role with the given name in a given project. Permissions are assigned to certain roles and through this to users.

We have updated the top of the projectroles documentation accordingly.

> This is great! Can you add 1-2 screenshots to the homepage/README? I think it is important to "hook" people in the README with a visual example, otherwise they may never even get to the "For the Impatient" section. Can you also add a link to the VarFish demo page directly on the README/homepage as a "See it in action"-type link?

Thank you for your suggestions. We have updated the README and manual front page accordingly.

We're looking forward to your further review.

Manuel
