### 2020-10-26 #1584 @holtgrewe
Dear all (in particular @olgabot and @arfon),

we're reallly sorry about the long delay but we got derailed by the usual and unusual things in academia...

Anyway, I think the main point to drive forward is:

- https://github.com/bihealth/sodar_core/issues/11

and our current planned change is

- https://github.com/bihealth/sodar_core/pull/14

There also is the raised point [#12](https://github.com/bihealth/sodar_core/issues/12) but I'd actually limit the documentation here as described in the ticket documentation.

We are looking forward to completing the review process.

Best wishes,
Manuel
