### 2020-11-04 #1584 @lpantano
Hey @rbatorsky, authors closed the issue you had opened, I don't see any more concern from you, let us know if we are missing something. If all sounds good, could you make sure you double check all the items under the list on the top of this issue, please? Thanks!
