### 2020-11-05 #1584 @olgabot
The most recent release is [v0.8.3](https://github.com/bihealth/sodar_core/releases/tag/v0.8.3) -- should this be updated in the issue?
