### 2020-11-05 bihealth/sodar_core#11 @olgabot
Thank you for addressing everything!
