### 2020-11-05 bihealth/sodar_core#11 @holtgrewe
Dear @olgabot - thanks! Sorry for taking so long to answer.
