### 2020-11-05 #1584 @lpantano
@rbatorsky, thank you! manual tests are ok as far as they work. In this case, probably it would be the example server. I think here, it would be this part: https://sodar-core.readthedocs.io/en/latest/for_the_impatient.html, if that tutorial works, I think we could check that box. Thank you!
