### 2020-11-05 #1584 @holtgrewe
I'm happy to clarify that there *are* automated unit tests:

- https://travis-ci.org/github/bihealth/sodar_core/branches

The `dev` branch has some issues and probably needs an update of the chrome driver (*sigh* the sad story of automated web testing these days...). However the `master` branch builds nicely.

Cheers,
Manuel
