### 2020-11-10 #1584 @lpantano
Hey @rbatorsky, could you quickly check that actually they have an automated test and it is working. Thanks!

https://travis-ci.org/github/bihealth/sodar_core/branches
