### 2020-11-11 #1584 @rbatorsky
Thanks @lpantano and sorry for the delay. Yes, I can verify that the manual tutorial steps do work and the automated tests seem to be passing. I also checked the version number since the github issue 0.6.2 matches the paper, but as was mentioned there have been later releases. I think my items are all set.
