### 2020-11-11 #1584 @lpantano
@holtgrewe, we are almost there, could you go over this list, please:

After that, I can then move forward with accepting the submission :)
