### 2020-11-12 #1584 @holtgrewe
Dear @lpantano, I'm happy to report that we made a v0.8.4 release of the software and deposited it here in Zenodo:

- https://zenodo.org/record/4269347#.X6z2564o-Ak

The DOI is `10.5281/zenodo.4269347` and we already updated the `paper.md` file in the sodar core repository to reflect that version update as well as the `codemeta.json` file.

Please tell us if we missed anything ;-)
